Refactor LastError as a public struct
